removed redundant ProjectGenerator class
